# ping/src/main/resources/openapi.yml
...
openapi: 3.1.0
paths:
...
  /api/projects:
    get:
      summary: List projects of user
      description: |
        List all the projects of the logged in user.

        The param onlyOwned specifies if the response should return all the projects the user is a member of,
        or only those the user owns.

        By default onlyOwned is considered as false.

        Any logged in user can access this endpoint.
      tags:
        - Project
      parameters:
        - name: onlyOwned
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of the projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectResponse'
        "403":
          description: Should not happen (this is automatically generated)
        "401":
          description: Not Authorized
      servers:
        - url: http://localhost:8080
      security:
        - SecurityScheme: []
    post:
      summary: Create a project
      description: |
        Create a new project with a name, the owner should also be in the list of members.

        Any logged in user can access this endpoint.
      tags:
        - Project
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewProjectRequest'
        required: true
      responses:
        "200":
          description: The created project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        "400":
          description: The project name is invalid (null or empty for example)
        "403":
          description: Should not happen (this is automatically generated)
        "401":
          description: Not Authorized
      servers:
        - url: http://localhost:8080
      security:
        - SecurityScheme: []
  /api/projects/all:
    get:
      summary: List all the projects
      description: |
        List all the projects available, be careful the response should not contain all the information about the members or the owner.

        Only an admin can access this endpoint.
      tags:
        - Project
      responses:
        "200":
          description: List of the projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectResponse'
        "403":
          description: The user is not an admin
        "401":
          description: Not Authorized
      servers:
        - url: http://localhost:8080
      security:
        - SecurityScheme:
            - admin
  /api/projects/{id}:
    put:
      summary: Update a project
      description: |
        Update a project, by changing its name or the owner of the project.

        If the new owner id or the new name are the same as the current one, it should not change anything and it is not an error.

        You should only update the fields that are not null in the request.
        If some fields are null, they should not be updated.

        If all the fields are null, it should return a bad request.

        Only the owner of the project or an admin can access this endpoint.
      tags:
        - Project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
        required: true
      responses:
        "200":
          description: The project was modified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        "404":
          description: "The project could not be found, or the new owner is not a\
            \ member of the project"
        "403":
          description: The user is not allowed to access the project
        "400":
          description: Both the name and the new owner are null
        "401":
          description: Not Authorized
      servers:
        - url: http://localhost:8080
      security:
        - SecurityScheme: []
    get:
      summary: Get a project
      description: |-
        Get a specific project based on its id.

        A user can access this endpoint if he is a member of the project or the owner.

        An admin can access this endpoint in any case.
      tags:
        - Project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "200":
          description: The project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        "404":
          description: Project not found
        "403":
          description: The user is not allowed to access the project
        "401":
          description: Not Authorized
      servers:
        - url: http://localhost:8080
      security:
        - SecurityScheme: []
    delete:
      summary: Delete a project
      description: |
        Delete a project.

        Only the owner of the project or an admin can access this endpoint.
      tags:
        - Project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "204":
          description: The project was deleted
        "404":
          description: The project could not be found
        "403":
          description: The user is not allowed to access this project
        "401":
          description: Not Authorized
      servers:
        - url: http://localhost:8080
      security:
        - SecurityScheme: []
  /api/projects/{id}/add-user:
    post:
      summary: Add a member
      description: |
        Add a member to a project.

        Any member of the project or an admin can access this endpoint.
      tags:
        - Project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProjectRequest'
        required: true
      responses:
        "204":
          description: The user was added to the project
        "404":
          description: "The project, or the user could not be found"
        "403":
          description: The user is not allowed to access the project
        "400":
          description: The userId is invalid (null or empty for example)
        "409":
          description: The user is already a member of the project
        "401":
          description: Not Authorized
      servers:
        - url: http://localhost:8080
      security:
        - SecurityScheme: []
  /api/projects/{id}/exec:
    post:
      summary: Execute a feature
      description: |
        Execute a feature on the project.

        For git features:

        If the project is not a git repository, it should return a bad request (except if the feature is git init).


        For git add, if a file does not exist it should return a bad request.

        Any other error should be returned as an internal server error.

        Any member of the project or an admin can access this endpoint.
      tags:
        - Project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecFeatureRequest'
        required: true
      responses:
        "204":
          description: The feature was successfully executed
        "404":
          description: The project could not be found
        "403":
          description: The user is not allowed to access the project
        "400":
          description: "Any parameter of the request is invalid (null or non-existent\
            \ feature)"
        "401":
          description: Not Authorized
      servers:
        - url: http://localhost:8080
      security:
        - SecurityScheme: []
  /api/projects/{id}/remove-user:
    post:
      summary: Remove a member
      description: |
        Remove a member from a project.

        Only the owner of the project or an admin can access this endpoint.
      tags:
        - Project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProjectRequest'
        required: true
      responses:
        "204":
          description: The user was removed from the project
        "404":
          description: "The project, or the user could not be found, the user to remove\
            \ is not a member of the project"
        "403":
          description: "The user is not allowed to access the project, or the user\
            \ to remove is the owner of the project"
        "400":
          description: The userId is invalid (null or empty for example)
        "401":
          description: Not Authorized
      servers:
        - url: http://localhost:8080
      security:
        - SecurityScheme: []
...
components:
  schemas:
    ExecFeatureRequest:
      type: object
      properties:
        feature:
          type: string
        command:
          type: string
        params:
          type: array
          items:
            type: string
...
    NewProjectRequest:
      type: object
      properties:
        name:
          type: string
...
    ProjectResponse:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
        members:
          type: array
          items:
            $ref: '#/components/schemas/UserSummaryResponse'
        owner:
          $ref: '#/components/schemas/UserSummaryResponse'
    UUID:
      type: string
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
        newOwnerId:
          $ref: '#/components/schemas/UUID'
...
    UserProjectRequest:
      type: object
      properties:
        userId:
          $ref: '#/components/schemas/UUID'
...
    UserSummaryResponse:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        displayName:
          type: string
        avatar:
          type: string
  securitySchemes:
    SecurityScheme:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authentication
...
info:
  title: ping API
  version: "1.0"